+++ backend/src/main/java/com/smartmealplanner/profile/persistence/DietaryPreference.java
@@ -30,7 +30,7 @@ public class DietaryPreference {
     private Boolean isExclusionary;
 
     @Column(name = "display_order", nullable = false)
-    private Integer displayOrder;
+    private Short displayOrder;
 
     @Column(name = "created_at", insertable = false, updatable = false, columnDefinition = "DATETIME(6)")
     private LocalDateTime createdAt;
@@ -46,13 +46,13 @@ public class DietaryPreference {
             String displayName,
             String description,
             Boolean isExclusionary,
-            Integer displayOrder) {
+            Short displayOrder) {
 
         this.code = code;
         this.displayName = displayName;
         this.description = description;
         this.isExclusionary = isExclusionary != null && isExclusionary;
-        this.displayOrder = displayOrder == null ? 1000 : displayOrder;
+        this.displayOrder = displayOrder == null ? (short) 1000 : displayOrder;
     }
 
     public Long id() {
@@ -75,7 +75,7 @@ public class DietaryPreference {
         return isExclusionary;
     }
 
-    public Integer displayOrder() {
+    public Short displayOrder() {
         return displayOrder;
     }
